Warn when app is not using the latest available version in their Ruby series
Developers should be using the latest Ruby version for their minor series. For example currently developers using 2.4.x should be using 2.4.5, if not then this PR will give them a warning:


```
There is a more recent Ruby version available for you to use:

ruby-2.4.5

The latest version will include security and bug fixes, we always recommend
running the latest version of your minor release.

See https://devcenter.heroku.com/articles/ruby-versions for all available versions.
```

This PR does not warn against EOL Ruby versions.

Upside: People don't know there are more recent Ruby versions available

Downside: This adds an extra second to a deploy that a user is not on the latest version (plus 1). For example if a user is using 2.5.0 and the latest is 2.5.5 then it effectively adds 6 seconds to their deploy. 

On one hand this is motivation to use the latest version and you get a free 5 seconds back! Alternatively I can spin up a thread and do this check in the background since the outcome does not affect.

For detecting EOL rubies we can either do a similar version traversal, but with some logic to account for major version bumps as well.