Move minor version checking logic to background
Since it is relatively expensive to check each version serially, we can instead queue up 5 version checks at a time using threads. If the last version check shows that it is a valid version, then it will enqueue another round of 5 checks in parallel.

This structure can be used from within the Ruby language pack to detect when a customer is not using the latest teeny/patch release for example if they're using 2.2.2 instead of 2.2.10.