Support use of BUNDLE_GEMFILE env var
This is to work around an issue with Bundler 1.15.2 where it cannot run successfully when the gemfile path is relative as described here: https://github.com/bundler/bundler/pull/5815

The Bundler 1.15.2 issue has been fixed in 1.16+, but the next supported version by the official ruby buildpack is Bundler 2 (starting with #v201) which not may not be feasible for an application to move to.

This issue does not affect running the app only the initial installation of gems.